fix: Ensure consistent data types in Gate class
Fixes a bug where objects could not be selected after loading a course file.

The `Gate` class constructor did not specify a `dtype` when creating its position and orientation numpy arrays. When loading from JSON, if the numbers had no decimal part, this could result in integer arrays. This caused issues with selection logic that implicitly expected floating-point data.

The `Gate` constructor now uses `dtype=float` to ensure consistency with the `Asteroid` class and prevent this bug.

+++ game_objects.py
@@ -84,8 +84,8 @@ class Gate:
     Represents a single race gate in the obstacle course.
     """
     def __init__(self, position, orientation, size=30):
-        self.position = np.array(position)
-        self.orientation = np.array(orientation)
+        self.position = np.array(position, dtype=float)
+        self.orientation = np.array(orientation, dtype=float)
         self.size = size # The radius of the gate opening
         self.is_passed = False
